fix moving file
